% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getLMS}
\alias{getLMS}
\alias{centilesLMS}
\title{Interpolate Centile or LMS tables}
\usage{
getLMS(
  x,
  sex,
  LMS,
  xname = "x",
  method = c("fmm", "periodic", "natural", "monoH.FC", "hyman", "linear", "constant"),
  rule = 1
)

centilesLMS(
  x,
  sex,
  LMS,
  cent = c(0.02, 0.05, 0.25, 0.5, 0.75, 0.95, 0.98),
  xname = "x",
  method = c("fmm", "periodic", "natural", "monoH.FC", "hyman", "linear", "constant"),
  rule = 1
)
}
\arguments{
\item{x}{a single value or vector of either ages in years or height in cm at which to interpolate LMS of centile values.}

\item{sex}{a single value or vector of sexes. Sex must be coded for males/females as 1/0, 1/2, 'm'/'f', 'M'/'F', 'Male'/'Female', or 'male'/'female'.}

\item{LMS}{a data frame containing the LMS table with columns named "Sex","Age","Lambda","Mu", "Sigma" or "Sex","Height","Lambda","Mu","Sigma". Internal LMS tables can be referenced by \code{pop.type}, i.e. \code{CDC.Height}}

\item{xname}{optional name to give the x column in the output, typically "Age" or "Height".}

\item{method}{the interpolation method. If it is "fmm", "periodic", "natural", "monoH.FC", or "hyman", \code{\link[stats]{splinefun}} is used. If \code{method} is "linear" or "constant", \code{\link[stats]{approxfun}} is used.}

\item{rule}{an integer (of length 1 or 2) describing how interpolation is to take place outside the interval [min(x), max(x)]. If \code{rule} is 1 then NAs are returned. If \code{rule} is 2 and method is "linear" or "constant", then \code{rule} is followed as in \code{\link[stats]{approxfun}}. If \code{rule} is 2 and a spline is used, the spline will be used to extrapolate outside the interval. If \code{rule} is length 2 (i.e c(2,1)), then different rules will be used for the left and right side extrapolation.}

\item{cent}{a single value or vector of centiles}
}
\value{
Returns a data frame with columns corresponding to the provided x-value and columns for either Lambda, Mu, and Sigma (for \code{getLMS}), or the requested centiles (for \code{centilesLMS}).
}
\description{
These functions are useful for producing reference tables or plots. \code{centilesLMS} produces a table of centiles. \code{getLMS} produces a table of interpolated LMS values.
}
\details{
\code{getLMS} produces a table of LMS values interpolated at the specified x values. Interpolation is done using either \code{\link[stats]{splinefun}} or \code{\link[stats]{approxfun}}, as determined by the \code{method} parameter. By default, observations outside the range of the reference LMS are returned as \code{NA}, but this can be overridden to allow for extrapolation by setting the \code{rule} parameter. See \code{\link[stats]{approxfun}}. \code{centilesLMS} uses \code{getLMS} to produce the appropriate LMS tabel, and then calculates the values at the centiles specified by \code{cent}
}
\examples{
#Produce a Tsimane height LMS table for males at ages 0-10
LMS<-getLMS(0:10,"Male",Tsimane.Height,xname="Age")

#Produce a Tsimane height centile table for males at ages 0-10
cent<-centilesLMS(0:10,"Male",Tsimane.Height,cent=c(0.025,0.5,0.975),xname="Age")

}
\seealso{
\code{\link{ZfromLMS}}, \code{\link{growthRef}}, \code{\link{getZ}}
}
