% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getZ}
\alias{getZ}
\alias{getCentile}
\alias{growthRef}
\title{Calculate Z-scores from growth references}
\usage{
getZ(
  x,
  y,
  sex,
  LMS,
  method = c("fmm", "periodic", "natural", "monoH.FC", "hyman", "linear", "constant"),
  rule = 1
)

getCentile(
  x,
  y,
  sex,
  LMS,
  method = c("fmm", "periodic", "natural", "monoH.FC", "hyman", "linear", "constant"),
  rule = 1
)

growthRef(
  x,
  y,
  sex,
  data,
  type = c("Weight", "Height", "BMI", "WFH"),
  pop = c("All", "CDC", "WHO", "Tsimane", "Shuar"),
  LMS = NULL,
  z = c("zscore", "centile"),
  method = c("fmm", "periodic", "natural", "monoH.FC", "hyman", "linear", "constant"),
  rule = 1
)
}
\arguments{
\item{x}{a single value or vector of either ages in years or height in cm at which to calcuate Z-scores, or the name of a column or item in \code{data}.}

\item{y}{a single value or vector of measures (heights, weights, BMIs) to be converted to Z-scores, or the name of a column or item in \code{data}. For included LMS references, height should be in cm, weight in kg, and BMI in kg/m^2.}

\item{sex}{a single value or vector of sexes, or the name of a column or item in \code{data}. Sex must be coded for males/females as 1/0, 1/2, 'm'/'f', 'M'/'F', 'Male'/'Female', or 'male'/'female'.}

\item{LMS}{a data frame containing the LMS table with columns named "Sex","Age","Lambda","Mu", "Sigma" or "Sex","Height","Lambda","Mu","Sigma". For \code{growthRef} this will typically be \code{NULL}, as \code{type} and \code{pop} will be used to determine the appropriate internal LMS tables.}

\item{method}{the interpolation method. If it is "fmm", "periodic", "natural", "monoH.FC", or "hyman", \code{\link[stats]{splinefun}} is used. If \code{method} is "linear" or "constant", \code{\link[stats]{approxfun}} is used.}

\item{rule}{an integer (of length 1 or 2) describing how interpolation is to take place outside the interval [min(x), max(x)]. If \code{rule} is 1 then NAs are returned. If \code{rule} is 2 and method is "linear" or "constant", then \code{rule} is followed as in \code{\link[stats]{approxfun}}. If \code{rule} is 2 and a spline is used, the spline will be used to extrapolate outside the interval. If \code{rule} is length 2 (i.e c(2,1)), then different rules will be used for the left and right side extrapolation.}

\item{data}{an optional data frame or list containing x, y, or sex.}

\item{type}{a string indicating the type of Z-score to calculate, corresponding to the measures in x and y. Either "Weight","Height","BMI", or "WFH", indicating weight-for-age, height-for-age, BMI-for-age, or weight-for-height.}

\item{pop}{the references or standards to return Z-scores based on. Can be "All","CDC","WHO","Tsimane",or "Shuar", with "All" equivalent to \code{pop=c("CDC","WHO","Tsimane","Shuar")}. Subsets are also allowed, i.e. \code{pop=c("CDC","Tsimane")}.}

\item{z}{a string. Either "zscore" or "centile" indicating the type of value to return.}
}
\value{
Returns either a data frame with columns corresponding to the requested Z-scores or centiles, or a vector, if only a single Z-score is requested.
}
\description{
Functions to calculates Z-scores from one or more growth references. \code{growthRef} provides an interface for indicating the desired LMS reference and passing growth data to \code{getZ}. \code{getZ} and \code{getCentile} calculate z-scores or centiles from child growth references provided in LMS format.
}
\details{
\code{growthRef} returns Z-scores calculated from lambda, mu, sigma (LMS) tables from various sources. \code{growthRef} acts as a wrapper that selects the appropriate LMS tables and passes them to \code{getZ} to interpolate LMS values and calculate z-scores at the exact age or height specified. At present, \code{growthRef} can provide z-scores based on four different growth references: the Tsimane, an indigenous group of Bolivia, the Shuar, an indigenous group of Ecuador, and references or standards from the US Center for Disease Control (CDC) and the World Health Organization (WHO). Internal LMS tables are typically referenced indirectly by \code{growthRef}, but can be referenced directly by \code{pop.type}, i.e. \code{Tsimane.Height}. See below for relevant citations and ranges for each growth reference.

\strong{Valid x-value ranges for included references and standards}
\tabular{lcccc}{
 \tab \strong{HFA (yrs)} \tab \strong{WFA (yrs)} \tab \strong{BFA (yrs)} \tab \strong{WFH (cm)} \cr
 \strong{CDC} \tab 0-20 \tab 0-20 \tab 2-20 \tab 45-121  \cr
 \strong{WHO} \tab 0-19 \tab 0-10 \tab 0-19 \tab 45-120 \cr
 \strong{Tsimane} \tab 0-25 \tab 0-25 \tab 0-25 \tab 45-170 \cr
 \strong{Shuar} \tab 0-25 \tab 0-25 \tab 0-25 \tab N/A
}

Note: For CDC and WHO references, height and length are combined together, with a junction at 2 years of age for height-for-age, and 80.5 cm for weight for height.

\code{getZ} calculates z-scores for growth data using an LMS data table, by first determining the appropriate L,M, and S values for a given observation by matching based on sex and then interpolating the L, M, and S values for the exact x-value (age or height) for the observation. Interpolation is done using either \code{\link[stats]{splinefun}} or \code{\link[stats]{approxfun}}, as determined by the \code{method} parameter. By default, observations outside the range of the reference LMS are returned as \code{NA}, but this can be overridden to allow for extrapolation by setting the \code{rule} parameter. See \code{\link[stats]{approxfun}}.

\code{getCentile} is simply a wrapper that passes parameters to \code{getZ} and then converts the z-scores to centiles using \code{pnorm}.
}
\section{References}{

\strong{Tsimane:} Blackwell, AD, Urlacher, SS, Beheim, B, von Rueden, C, Jaeggi, A, Stieglitz, J, Tumble, BC, Gurven, MD, Kaplan, H. (2017) Growth references for Tsimane forager-horticulturalists of the Bolivian Amazon, American Journal of Physical Anthropology. 162(3) 441-461. DOI: 10.1002/ajpa.23128.

\strong{Shuar:} Urlacher, SS, Blackwell, AD, Liebert, MA, Madimenos, FC, Cepon-Robins, TJ, Gildner, TE, Snodgrass, JJ, Sugiyama, LS. (2016) Physical Growth of the Shuar: Height, Weight, and BMI Growth References for an Indigenous Amazonian Population, American Journal of Human Biology, 28(1): 16-30. DOI: 10.1002/ajhb.22747.

\strong{CDC:} Kuczmarski, R. J., Ogden, C. L., Guo, S. S., Grummer-Strawn, L. M., Flegal, K. M., Mei, Z., Wei, R., Curtin, L. R., Roche, A. F., Johnson, C. L. (2002). 2000 CDC Growth Charts for the United States: Methods and development. Vital Health Statistics 11,1-190.

\url{https://www.cdc.gov/growthcharts/data_tables.htm}

\strong{WHO:} World Health Organization. (2006). New child growth standards: Length/ height-for-age, weight-for-age, weight-for-length, weight-for-height and body mass index-for-age: Methods and development. Geneva: World Health Organization

de Onis, M., Onyango, A., Borghi, E., Siyam, A., Nishida, C., & Siekman, J. (2007). Development of a WHO growth reference for school-aged children and adolescents. Bulletin of World Health Organization, 85, 660-667.

\url{http://www.who.int/childgrowth/en/}. \url{http://www.who.int/growthref/en/}.
}

\examples{
#Height-for-age
data(TsimaneData)
HFA<-growthRef(Age,Height,Sex,data=TsimaneData,type="Height")

#Height-for-age, just CDC and Tsimane
HFA<-growthRef(Age,Height,Sex,data=TsimaneData,type="Height",pop=c("CDC","Tsimane"))

#Weight-for-height, CDC centiles only
WFH<-growthRef(Height,Weight,Sex,data=TsimaneData,type="WFH",pop="CDC",z="centile")

#The same as above, but referencing the CDC.WFH table directly
WFH<-growthRef(Height,Weight,Sex,data=TsimaneData,LMS=CDC.WFH,z="centile")

#Tsimane specific z-score for a single individual, age 2 yrs, 75 cm, male.
getZ(2,75,"Male",Tsimane.Height)

LMS<-getLMS(0:10,"Male",Tsimane.Height)
cent<-centilesLMS(0:10,"Male",Tsimane.Height,cent=c(0.025,0.5,0.975))
}
\seealso{
\code{\link{ZfromLMS}}, \code{\link{splinefun}}, \code{\link{approxfun}}
}
