% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ZfromLMS}
\alias{ZfromLMS}
\title{Calculate z-scores}
\usage{
ZfromLMS(measure, la, mu, si)
}
\arguments{
\item{measure}{a vector of numbers}

\item{la}{a vector of lambda values}

\item{mu}{a vector of mu values}

\item{si}{a vector of sigma values}
}
\description{
Calculates z-scores from a measure and LMS values
}
\details{
\code{ZfromLMS} calcuates a z-score from a \code{measure}=\eqn{y} and the distribuiton parameters, \code{la}=\eqn{\lambda}, \code{mu}=\eqn{\mu}, and \code{si}=\eqn{\sigma}, which describe the median, the generalized coefficient of variation, and the power in the Box-Cox transformation. It uses the formula: \eqn{Z=((y / \mu)^\lambda - 1)/(\lambda * \sigma)}
}
